fix method_not_allowed and not_found for python3
The body of a response needs to be bytes. Assuming we are in relatively modern Python
then we can make this small change to get 404 and 405 working in an environment that
doesn't already have some kind of middleware in place to turn strings into bytes.

In my previous efforts with Selector I did have such a tool in place and so never noticed
this. I'm now using Selector in a small demo and stumbled on this.